+++ impala/sqlalchemy.py
@@ -24,8 +24,9 @@ from sqlalchemy.engine.default import DefaultDialect, DefaultExecutionContext
 from sqlalchemy.sql.compiler import (DDLCompiler, GenericTypeCompiler,
                                      IdentifierPreparer)
 from sqlalchemy.types import (BOOLEAN, SMALLINT, BIGINT, TIMESTAMP, FLOAT,
-                              DECIMAL, Integer, Float, String, VARCHAR,
+                              DECIMAL, Integer, Float, String, CHAR, VARCHAR,
                               DATE)
+from sqlalchemy import text
 
 
 registry.register('impala', 'impala.sqlalchemy', 'ImpalaDialect')
@@ -170,6 +171,7 @@ _impala_type_to_sqlalchemy_type = {
     'DOUBLE': DOUBLE,
     'STRING': STRING,
     'DECIMAL': DECIMAL,
+    'CHAR': CHAR,
     'VARCHAR': VARCHAR,
     'DATE': DATE
     }
@@ -222,9 +224,9 @@ class ImpalaDialect(DefaultDialect):
         self.default_schema_name = connection.connection.default_db
 
     def _get_server_version_info(self, connection):
-        raw = connection.execute('select version()').scalar()
+        raw = connection.execute(text('select version()')).scalar()
         v = raw.split()[2]
-        m = re.match('.*?(\d{1,3})\.(\d{1,3})\.(\d{1,3}).*', v)
+        m = re.match(r'.*?(\d{1,3})\.(\d{1,3})\.(\d{1,3}).*', v)
         return tuple([int(x) for x in m.group(1, 2, 3) if x is not None])
 
     def has_table(self, connection, table_name, schema=None):
@@ -241,12 +243,17 @@ class ImpalaDialect(DefaultDialect):
             query += ' IN %s' % escaped_schema
         tables = [
             tup[1] if len(tup) > 1 else tup[0]
-            for tup in connection.execute(query).fetchall()
+            for tup in connection.execute(text(query)).fetchall()
         ]
         return tables
 
+    def get_view_names(self, connection, schema=None, **kw):
+        # Impala doesn't distinguish between tables and view when calling
+        # SHOW TABLES. So return a blank list.
+        return []
+
     def get_schema_names(self, connection, **kw):
-        rp = connection.execute("SHOW SCHEMAS")
+        rp = connection.execute(text("SHOW SCHEMAS"))
         return [r[0] for r in rp]
 
     def get_columns(self, connection, table_name, schema=None, **kwargs):
@@ -255,15 +262,16 @@ class ImpalaDialect(DefaultDialect):
         if schema is not None:
             name = '%s.%s' % (schema, name)
         query = 'SELECT * FROM %s LIMIT 0' % name
-        cursor = connection.execute(query)
+        cursor = connection.execute(text(query))
         schema = cursor.cursor.description
         # We need to fetch the empty results otherwise these queries remain in
         # flight
         cursor.fetchall()
         column_info = []
+        # `select * from table` results in 'tablename.columnname', so strip off 'tablename.'
         for col in schema:
             column_info.append({
-                'name': col[0],
+                'name': col[0].split('.')[-1].strip() if '.' in col[0] else col[0],
                 'type': _impala_type_to_sqlalchemy_type[col[1]],
                 'nullable': True,
                 'autoincrement': False})