+++ impala/util.py
@@ -14,6 +14,7 @@
 
 from __future__ import absolute_import
 
+import sys
 import warnings
 import logging
 import string
@@ -167,12 +168,6 @@ def warn_deprecate(functionality='This', alternative=None):
     warnings.warn(msg, Warning)
 
 
-def warn_unused_jwt():
-    msg = ("The JWT argument is ignored as auth_mechanism is not 'JWT'. "
-           "Specify auth_mechanism='JWT' to make use of JWT authentication.")
-    warnings.warn(msg, Warning)
-
-
 def warn_nontls_jwt():
     msg = ("JWT authentication is running without SSL/TLS. This is not a secure "
            "configuration unless other layers are providing transport security.")
@@ -207,14 +202,47 @@ def get_cookie_expiry(c):
     return None
 
 
-def get_all_matching_cookies(cookie_names, path, resp_headers):
+def get_cookies(resp_headers):
+    """Returns a SimpleCookie containing all Set-Cookie entries in resp_headers."""
     if 'Set-Cookie' not in resp_headers:
         return None
 
     cookies = http_cookies.SimpleCookie()
     try:
-        cookies.load(resp_headers['Set-Cookie'])
-    except:
+        if sys.version_info.major == 2:
+            cookies.load(resp_headers['Set-Cookie'])
+        else:
+            cookie_headers = resp_headers.get_all('Set-Cookie')
+            for header in cookie_headers:
+                cookies.load(header)
+        return cookies
+    except Exception:
+        return None
+
+
+def get_all_cookies(path, resp_headers):
+    """Return cookies that match path.
+
+    Returns a list of Morsel objects representing cookie key/value pairs for all cookies
+    in resp_headers matching path."""
+    cookies = get_cookies(resp_headers)
+    if not cookies:
+        return None
+
+    matching_cookies = []
+    for c in cookies.values():
+        if c and cookie_matches_path(c, path):
+            matching_cookies.append(c)
+    return matching_cookies
+
+
+def get_all_matching_cookies(cookie_names, path, resp_headers):
+    """Return cookies in cookie_names that match path.
+
+    Returns a list of Morsel objects representing cookie key/value pairs for cookies
+    in resp_headers matching path where the cookie is also listed in cookie_names."""
+    cookies = get_cookies(resp_headers)
+    if not cookies:
         return None
 
     matching_cookies = []